+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Tous les changements de ce projet seront documentés dans ce fichier.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).*
  
+## 03-01-2022
+## Added 
+- Ajout de la class Game ainsi que son controller
+
+
 ## Amélioration possible ? 
 - Meilleur agencement des boutons pour le modal